wip: migrate overviews to third-level dropdown
