+++ apps/frontend-manage/src/components/courses/LiveQuizList.tsx
@@ -21,14 +21,14 @@ function LiveQuizList({
         <QRCodePopover
           triggerStyle="basic"
           triggerText={t('manage.general.qrCode')}
-          relHref={`/course/${courseId}/liveQuizzes`}
+          relHref={`/course/${courseId}/liveQuizzes/overview`}
           data={{ cy: `qr-link-live-quiz-list` }}
         />
         <Button
           basic
           onClick={async () => {
             try {
-              const link = `${process.env.NEXT_PUBLIC_LTI_URL}?redirectTo=${process.env.NEXT_PUBLIC_PWA_URL}/course/${courseId}/liveQuizzes`
+              const link = `${process.env.NEXT_PUBLIC_LTI_URL}?redirectTo=${process.env.NEXT_PUBLIC_PWA_URL}/course/${courseId}/liveQuizzes/overview`
               console.log(link)
               await navigator.clipboard.writeText(link)
             } catch (e) {
