+++ apps/frontend-pwa/src/pages/course/[courseId]/liveQuizzes/[id].tsx
